+++ src/Straumr.Cli/Commands/Request/RequestDeleteCommand.cs
@@ -1,15 +1,25 @@
 using Spectre.Console;
 using Spectre.Console.Cli;
+using Straumr.Core.Enums;
+using Straumr.Core.Exceptions;
 using Straumr.Core.Services.Interfaces;
 
 namespace Straumr.Cli.Commands.Request;
 
-public class RequestDeleteCommand(IStraumrRequestService requestService)
+public class RequestDeleteCommand(IStraumrOptionsService optionsService, IStraumrRequestService requestService)
     : AsyncCommand<RequestDeleteCommand.Settings>
 {
     public override async Task<int> ExecuteAsync(CommandContext context, Settings settings,
         CancellationToken cancellation)
     {
+        bool hasWorkspace = optionsService.Options.CurrentWorkspace != null;
+
+        if (!hasWorkspace)
+        {
+            throw new StraumrException("No workspace loaded. Please load a workspace using 'workspace use <name>'",
+                StraumrError.MissingEntry);
+        }
+
         await requestService.DeleteAsync(settings.Identifier);
         AnsiConsole.MarkupLine($"[red]Deleted request[/] [bold]{settings.Identifier}[/]");
         return 0;
